#2517 divide `MasterWindow` into data layer, and presentational component
